// src/modules/role/domain/enums/role-reference.enum.ts
export enum ERoleReference {
  ADMIN = 'ADMIN',
  ADMIN_ORGANIZATION = 'ADMIN_ORGANIZATION',
  ADMIN_COMPANY = 'ADMIN_COMPANY',
  CLIENT = 'CLIENT',
}

export enum ERoleReferenceTranslate {
  ADMIN = 'Admin',
  ADMIN_ORGANIZATION = 'Admin. Organização',
  ADMIN_COMPANY = 'Admin. Empresa',
  CLIENT = 'Cliente',
}